Need to also rename app_config.php -> AppConfig.php, for consistency with the class name.
