+++ .github/actions/web_config/action.yml
@@ -1,6 +1,6 @@
 name: "Set up config file for a run"
 author: Luke Swanson
-description: "Given a set of appropriate inputs, this action prepares the app_config.php file for a website deploy."
+description: "Given a set of appropriate inputs, this action prepares the AppConfig.php file for a website deploy."
 branding:
   icon: 'clipboard'
   color: 'blue'
@@ -13,5 +13,5 @@ runs:
   using: "composite"
   steps:
     - name: Set API Base URL
-      run: sed -i 's@$APP_CONFIG\['\''WEBSITE_API_URL_BASE'\''\]\s*=\s*'\''.*'\'';@$APP_CONFIG['\''WEBSITE_API_URL_BASE'\''] = '\''${{inputs.api_base}}'\'';@g' site/config/app_config.php
+      run: sed -i 's@$APP_CONFIG\['\''WEBSITE_API_URL_BASE'\''\]\s*=\s*'\''.*'\'';@$APP_CONFIG['\''WEBSITE_API_URL_BASE'\''] = '\''${{inputs.api_base}}'\'';@g' site/config/AppConfig.php
       shell: bash
